+++ tests/testthat/test_log.R
@@ -34,6 +34,17 @@ test_that('alt base', {
   expect_equal(rec_trans, exp_res)
 })
 
+test_that('alt offset', {
+  rec <- recipe(~., data = ex_dat) %>%
+    step_log(x1, x2, x3, x4, base = pi, offset = 0.1)
+  
+  rec_trained <- prep(rec, training = ex_dat, verbose = FALSE)
+  rec_trans <- bake(rec_trained, newdata = ex_dat)
+  
+  exp_res <- as_tibble(lapply(ex_dat, function(x) log(x + 0.1, base = pi)))
+  
+  expect_equal(rec_trans, exp_res)
+})
 
 test_that('printing', {
   rec <- recipe(~., data = ex_dat) %>%